Working demo now functional
Added a working demo , updated the binder function , two default shaders
added that perform very basic tasks but need updating , vbo slightly
modified to handle rotation and pass on translation and the example
class now works. Inputhandler is a new class that function and allows
input to be taken for a large number of keys . Most of this still needs
testing.